requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
mysql-connector-python==8.0.33
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.1
google-api-python-client==2.92.0
